# llm-tools/custom-gpt/actions/openapi.yaml
openapi: 3.1.0
info:
  title: Professional Profile Analyser API
  version: 1.0.0
  description: |
    JSON API for Custom GPT Actions. System of record for jobs, candidates,
    evaluations, and Compensation Intelligence research. LLM interprets;
    this API decides and persists scores / returns market-pay evidence.

    Base URL is deployment-specific. Use a publicly reachable HTTPS origin
    when configuring ChatGPT Actions (e.g. tunnel to local/dev).
servers:
  - url: https://YOUR_PPA_HOST
    description: Replace with the deployed Professional Profile Analyser base URL
security:
  - ApiKeyAuth: []
  - BearerAuth: []
...
paths:
  /api/gpt/jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List jobs
      description: |
        List persisted jobs. Optional `q` filters by title, description, profile, or id.
        Use before answering questions about stored opportunities.
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
          description: Case-insensitive search text
      responses:
        "200":
          description: Job list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/gpt/jobs/{job_id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get job
      description: |
        Retrieve a job including analysis (role intent, must-haves, skill tiers).
        Prefer this over conversational memory for stored jobs.
      parameters:
        - $ref: "#/components/parameters/jobId"
      responses:
        "200":
          description: Job detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/gpt/candidates:
    get:
      operationId: listCandidates
      tags: [candidates]
      summary: List candidates
      description: |
        List candidate index entries without full resume text.
        Optional `q` filters by name, role, city, or id.
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Candidate list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandidateListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/gpt/candidates/{candidate_id}:
    get:
      operationId: getCandidate
      tags: [candidates]
      summary: Get candidate
      description: |
        Retrieve a candidate. Set include_resume=true only when evidence inspection
        is required. Never invent candidates.
      parameters:
        - $ref: "#/components/parameters/candidateId"
        - name: include_resume
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: Include full resume_text when true
      responses:
        "200":
          description: Candidate detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandidateDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/gpt/jobs/{job_id}/candidates:
    get:
      operationId: listJobCandidates
      tags: [candidates]
      summary: List candidates evaluated for a job
      description: Return candidates that have a persisted evaluation for the given job.
      parameters:
        - $ref: "#/components/parameters/jobId"
      responses:
        "200":
          description: Candidates with evaluation summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobCandidatesResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/gpt/evaluations:
    get:
      operationId: listEvaluations
      tags: [evaluations]
      summary: List evaluations
      description: List persisted evaluations. Filter by job_id and/or candidate_id.
      parameters:
        - name: job_id
          in: query
          required: false
          schema:
            type: string
        - name: candidate_id
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Evaluation list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvaluationListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: evaluateCandidate
      tags: [evaluations]
      summary: Create or re-evaluate candidate against job
      description: |
        WRITE operation. Starts async evaluation via the application scoring engine
        (default v2). Returns a Task — poll getTask. Execute only with clear user intent
        to evaluate or re-evaluate.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvaluateRequest"
      responses:
        "200":
          description: Async task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/gpt/jobs/{job_id}/evaluations:
    get:
      operationId: listJobEvaluations
      tags: [evaluations]
      summary: List evaluations for a job
      parameters:
        - $ref: "#/components/parameters/jobId"
      responses:
        "200":
          description: Job evaluations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobEvaluationsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/gpt/evaluations/{evaluation_id}:
    get:
      operationId: getEvaluation
      tags: [evaluations]
      summary: Get evaluation
      description: |
        Retrieve persisted evaluation by id (`{job_id}_{candidate_id}`).
        Treat final_score as authoritative — do not recalculate.
        Use include_items=true for skill-level evidence.
      parameters:
        - $ref: "#/components/parameters/evaluationId"
        - name: include_items
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Evaluation detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvaluationDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/gpt/jobs/{job_id}/evaluations/{candidate_id}:
    get:
      operationId: getJobCandidateEvaluation
      tags: [evaluations]
      summary: Get evaluation by job and candidate
      parameters:
        - $ref: "#/components/parameters/jobId"
        - $ref: "#/components/parameters/candidateId"
        - name: include_items
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Evaluation detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvaluationDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/gpt/compensation/prefill/{job_id}:
    get:
      operationId: prefillCompensationFromJob
      tags: [compensation]
      summary: Prefill compensation research from a job
      description: |
        Map a stored job into a CompensationResearchRequest-shaped prefill
        (profile, skills, seniority, allocation, contract, location, source_job_id).
        Prefer this when discussing an open job already in the system.
      parameters:
        - $ref: "#/components/parameters/jobId"
      responses:
        "200":
          description: Prefill payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompensationPrefill"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/gpt/compensation/history:
    get:
      operationId: listCompensationHistory
      tags: [compensation]
      summary: List cached compensation research
      description: |
        Compact summaries of cached Compensation Intelligence runs.
        Use before re-running research when a recent cache may suffice.
      responses:
        "200":
          description: History list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompensationHistoryListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/gpt/compensation/history/{cache_key}:
    get:
      operationId: getCompensationHistoryItem
      tags: [compensation]
      summary: Get cached compensation research
      description: |
        Load cached research by cache_key. Observations omitted by default
        (`observations_omitted=true`); set include_observations=true to cite rows.
      parameters:
        - name: cache_key
          in: path
          required: true
          schema:
            type: string
        - name: include_observations
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Cached research payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompensationResearchResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/gpt/compensation/research/wait:
    post:
      operationId: researchCompensationWait
      tags: [compensation]
      summary: Run compensation research and wait for the final result
      description: |
        Convenience for clients that cannot poll Tasks reliably (e.g. Open WebUI tools).
        Runs research and returns the final payload in one call.
        Cache policy: force_refresh defaults to false — use cache unless the user
        explicitly asks to ignore cache / force a new research.
        Prefer researchCompensationAsync + getTask when polling works.
        Never invent salaries — cite only returned evidence.
      parameters:
        - name: include_observations
          in: query
          required: false
          schema:
            type: boolean
            default: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompensationResearchRequest"
      responses:
        "200":
          description: Research result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompensationResearchResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          description: Validation error
  /api/gpt/compensation/research/async:
    post:
      operationId: researchCompensationAsync
      tags: [compensation]
      summary: Start async compensation / market-pay research
      description: |
        Preferred GPT path for long-running Compensation Intelligence research.
        Returns a Task (`kind=compensation_research`). Poll getTask; on success
        use `task.result`. Never invent salaries — cite only returned evidence.
        Execute only with clear user intent to research market pay.
        Cache policy: force_refresh defaults to false; set true only when the user
        explicitly asks to ignore cache / force a new research.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompensationResearchRequest"
      responses:
        "200":
          description: Async task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          description: Validation error
...
  /api/gpt/tasks/{task_id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Poll async task
      description: |
        Poll analyse/extract/score/compensation_research task status under the gpt resource.
        When completed after evaluateCandidate, fetch the evaluation via getJobCandidateEvaluation.
        When completed after researchCompensationAsync, use task.result as the compensation payload.
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Task status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Set to ${PROFESSIONAL_PROFILE_API_KEY} in GPT Builder (never commit real keys).
    BearerAuth:
      type: http
      scheme: bearer
      description: Authorization Bearer ${PROFESSIONAL_PROFILE_API_KEY}
  parameters:
    jobId:
      name: job_id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    JobDetail:
      allOf:
        - $ref: "#/components/schemas/JobSummary"
        - type: object
          properties:
            job_description: { type: string }
            ideal_candidate_context: { type: string }
            analysis: { $ref: "#/components/schemas/JobAnalysis" }
    JobListResponse:
      type: object
      properties:
        count: { type: integer }
        jobs:
          type: array
          items: { $ref: "#/components/schemas/JobSummary" }
...
    CandidateListResponse:
      type: object
      properties:
        count: { type: integer }
        candidates:
          type: array
          items: { $ref: "#/components/schemas/CandidateSummary" }
...